Scope all video controller actions to current user.
Override action to pass current_user as a parameter.
Modify actions to accept the parameter and use it.
For edits, build an association and put it in the changeset.
For views, use assoc() to scope the query to the current user. Use the
query to control what's retrieved from the database.