Annotate videos.
Remove ping functionality.

Update video.js to push a message to the channel when an annotation is
posted. Render an annotation when the message is received. This can be
from other clients too.

Update Video channel to broadcast the new_annotation event to other
clients when it receives the message from a client.