Handle temporary channel disconnects.
Keep track of the last seen annotation ID so they don't show up twice in
the msgContainer when the server goes away and comes back later.

Modify the client to track the last seen ID in the channel params
object. This is sent to the server on join.

Modify the server video channel to use the last seen ID if available to
filter the annotations to return to the client.